monc: two new functions to make handling subscriptions easier
First, let a client query for its current subscription state. Right now
this is just the requested version (no way to look at the flags).
Second, provide a subscription-modifying function which will not decrement
the start version.
These are tailored pretty narrowly for the current needs of the OSD, but
should extend cleanly.

Signed-off-by: Greg Farnum <greg@inktank.com>